unified configuration
