+++ main.go
@@ -13,26 +13,14 @@ import (
 	"k8s.io/klog/v2"
 )
 
-var (
-	configFilePath               string
-	resyncInterval               uint
-	eventListenerType            string
-	stateKey                     string
-	deleteDependents             bool
-	createMissingRelatedEntities bool
-	portBaseURL                  string
-	portClientId                 string
-	portClientSecret             string
-)
-
 func initiateHandler(exporterConfig *port.Config, k8sClient *k8s.Client, portClient *cli.PortClient) (*handlers.ControllersHandler, error) {
-	apiConfig, err := integration.GetIntegrationConfig(portClient, stateKey)
+	apiConfig, err := integration.GetIntegrationConfig(portClient, config.ApplicationConfig.StateKey)
 	if err != nil {
 		klog.Fatalf("Error getting K8s integration config: %s", err.Error())
 	}
 
-	cli.WithDeleteDependents(deleteDependents)(portClient)
-	cli.WithCreateMissingRelatedEntities(createMissingRelatedEntities)(portClient)
+	cli.WithDeleteDependents(apiConfig.DeleteDependents)(portClient)
+	cli.WithCreateMissingRelatedEntities(apiConfig.CreateMissingRelatedEntities)(portClient)
 
 	newHandler := handlers.NewControllersHandler(exporterConfig, apiConfig, k8sClient, portClient)
 	newHandler.Handle()
@@ -56,21 +44,21 @@ func main() {
 		klog.Fatalf("Error building K8s client: %s", err.Error())
 	}
 
-	portClient, err := cli.New(portBaseURL,
-		cli.WithClientID(portClientId), cli.WithClientSecret(portClientSecret),
-		cli.WithHeader("User-Agent", fmt.Sprintf("port-k8s-exporter/0.1 (statekey/%s)", stateKey)),
+	portClient, err := cli.New(config.ApplicationConfig.PortBaseURL,
+		cli.WithClientID(config.ApplicationConfig.PortClientId), cli.WithClientSecret(config.ApplicationConfig.PortClientSecret),
+		cli.WithHeader("User-Agent", fmt.Sprintf("port-k8s-exporter/0.1 (statekey/%s)", config.ApplicationConfig.StateKey)),
 	)
 
 	if err != nil {
 		klog.Fatalf("Error building Port client: %s", err.Error())
 	}
 
-	exporterConfig, err := config.GetConfigFile(configFilePath, resyncInterval, stateKey, eventListenerType)
+	exporterConfig, err := config.GetConfigFile(config.ApplicationConfig.ConfigFilePath, config.ApplicationConfig.ResyncInterval, config.ApplicationConfig.StateKey, config.ApplicationConfig.EventListenerType)
 	if err != nil {
 		klog.Fatalf("Error building Port K8s Exporter config: %s", err.Error())
 	}
 
-	_, err = integration.GetIntegrationConfig(portClient, stateKey)
+	_, err = integration.GetIntegrationConfig(portClient, config.ApplicationConfig.StateKey)
 	if err != nil {
 		if exporterConfig == nil {
 			klog.Fatalf("The integration does not exist and no config file was provided")
@@ -83,7 +71,7 @@ func main() {
 
 	klog.Info("Starting controllers handler")
 	handler, _ := initiateHandler(exporterConfig, k8sClient, portClient)
-	eventListener := event_listener.NewEventListener(stateKey, eventListenerType, handler, portClient)
+	eventListener := event_listener.NewEventListener(config.ApplicationConfig.StateKey, config.ApplicationConfig.EventListenerType, handler, portClient)
 	err = eventListener.Start(func(handler *handlers.ControllersHandler) (*handlers.ControllersHandler, error) {
 		handler.Stop()
 		return initiateHandler(exporterConfig, k8sClient, portClient)
@@ -94,14 +82,5 @@ func main() {
 }
 
 func init() {
-	configFilePath = config.NewString("config", "", "Path to Port K8s Exporter config file. Required.")
-	stateKey = config.NewString("state-key", "", "Port K8s Exporter state key id. Required.")
 
-	resyncInterval = config.NewUInt("resync-interval", 0, "The re-sync interval in minutes. Optional.")
-	portBaseURL = config.NewString("port-base-url", "https://api.getport.io", "Port base URL. Optional.")
-
-	portClientId = config.NewString("port-client-id", "", "Port client id. Required.")
-	portClientSecret = config.NewString("port-client-secret", "", "Port client secret. Required.")
-
-	eventListenerType = config.NewString("event-listener-type", "POLLING", "Event listener type. Optional.")
 }
